refactor dashboard code
